+++ app/Http/Requests/V1/StoreCustomerRequest.php
@@ -1,0 +1,43 @@
+<?php
+
+namespace App\Http\Requests\V1;
+
+use Illuminate\Foundation\Http\FormRequest;
+use Illuminate\Validation\Rule;
+
+class StoreCustomerRequest extends FormRequest
+{
+    /**
+     * Determine if the user is authorized to make this request.
+     */
+    public function authorize(): bool
+    {
+        // TODO: implement authorization
+        return true;
+    }
+
+    /**
+     * Get the validation rules that apply to the request.
+     *
+     * @return array<string, \Illuminate\Contracts\Validation\Rule|array|string>
+     */
+    public function rules(): array
+    {
+        return [
+            'name'       => ['required'],
+            'type'       => ['required', Rule::in('I', 'B', 'i', 'b')],
+            'email'      => ['required', 'email'],
+            'address'    => ['required'],
+            'city'       => ['required'],
+            'state'      => ['required'],
+            'postalCode' => ['required']
+        ];
+    }
+
+    protected function prepareForValidation()
+    {
+        $this->merge([
+            'postal_code' => $this->postalCode
+        ]);
+    }
+}
